3054 : Resource existing in 2 directories

+++ Dev/Dev2.Studio.Core/Workspaces/WorkspaceItemRepository.cs
@@ -217,7 +217,10 @@ namespace Dev2.Workspaces
 
             WorkspaceItems.Remove(itemToRemove);
             Write();
-            resourceModel.Environment.ResourceRepository.DeleteResourceFromWorkspaceAsync(resourceModel);
+            if (!resourceModel.IsNotWarewolfPath)
+            {
+                resourceModel.Environment.ResourceRepository.DeleteResourceFromWorkspaceAsync(resourceModel);
+            }
         }
 
         public void ClearWorkspaceItems(IContextualResourceModel resourceModel)
